Adds a way to view comments to Core as well as a time zone feature for games.
Times stored in the database are stored in the database as inputted to the web form.
Although the user intends the "game start", "game end", and other times to be in their
local timezone, they are read out of the database in UTC. Thus, all times are actually
5 hours ahead (for EST, that is).
So, when comparing times with the current time (i.e. to check if registration has begun),
do the following:

  if (Time.now + @current_game.utc_offset > @current_game.registration_begins)