Restrict heatmap to admins for now. Obfuscate the data.
We'll release the heatmap to everyone later. For now, just admins.

Also it might be possible to see where the "hiding spot du jour" is by
looking at the last tag's LatLng. This sorts all the tags by latitude so
it least requires a bit more cleverness to find out where the last tag
occurred.