Refactor some waiver things
The whole registration process sucks - it's flaky and bad. This is the
first step of refactoring it (it probably breaks it actually, I'm not
testing this stuff live).

The only db change here is moving the `dateofbirth` column to the Person
model, which makes more sense because it doesn't vary from year-to-year
and allows us a cleaner way of determining whether a Person can legally
sign a waiver or not.