Remove caching and actually include needed associations
So, this improves GamesController#show's rendering speed quite a bit.
The N+1's are gone, and so is the caching. On my laptop, rendering 500
players takes about 5 seconds, with about 3.4 seconds in views and 1.6
getting lost in the rails ether (only 23.7ms in ActiveRecord).

Maybe a bit too slow to remain without caching. This uses
ActionController::ConditionalGet#stale? to return a cached version of a
page if the Game instance hasn't changed. Although the game instance
changes every minute during the game, maybe nginx as a reverse proxy can
roll together everyone's requests...?