Asset Pipeline and Autocomplete refactor
So I wanted to rip out jQuery UI because it's big and only used for the
autocomplete box. So then I ripped out the autocomplete box and built a
new one with Backbone. But then everything is bad because it's in
public/ and doesn't use the assets pipeline. So then I moved everything
around and probably broke things in a bad way.

Anyway, here it is.