Lots more speed improvements for scoreboard page
Fix a lot more N+1's by including literally everything when I fetch the
registrations. This is still way overdoing it because not all that
information is necessary but this improved the speed about 3x over the
last commit.

Total speed improvements from this series of fixes totals somewhere
around a 10x improvement.

Also committed the tools show page which I forgot in the last commit.